Move tests to a new package
