+++ src/test/java/com/checkmarx/flow/cucumber/integration/sca_scanner/bugtrackers/ado/ScaIssuesCreationRunner.java
@@ -1,4 +1,4 @@
-package com.checkmarx.flow.cucumber.integration.azure.publishing.issueprocessing.scaissueprocessing;
+package com.checkmarx.flow.cucumber.integration.sca_scanner.bugtrackers.ado;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
@@ -7,7 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = "src/test/resources/cucumber/features/integrationTests/sca/scanResultsProcessing.feature",
-        glue = {"com.checkmarx.flow.cucumber.integration.azure.publishing.issueprocessing.scaissueprocessing"},
+        glue = {"com.checkmarx.flow.cucumber.integration.sca_scanner.bugtrackers.ado"},
         tags = "@SCA_Issues_Creation and not @Skip")
 public class ScaIssuesCreationRunner {
 }
